cleanup

+++ src/parse/parse.h
@@ -1,7 +1,10 @@
 #pragma once
 
-#include "lem_in.h"
-#include "room_map.h"
+#include "allocator/safe_malloc.h"
+#include "parsing_iterator.h"
+#include "readfile/readfile.h"
+#include "containers/maps.h"
+#include "containers/vec.h"
 
 // 0   == StartRoom
 // n-1 == EndRoom
@@ -13,11 +16,11 @@ typedef struct {
 	t_room		**rooms;
 	/* (name -> room) */
 	room_map	rooms_map;
-	uint		nb_ants;
-	uint		nb_rooms;
+	u64			nb_ants;
+	u64			nb_rooms;
 
 	/* Private */
-	uint		_indexing;
+	u64			_indexing;
 	t_room		*_start_room;
 	t_room		*_end_room;
 }	t_farm;
